requirements.txt
# openlens Requirements
# Python 3.6+ required

# GUI Framework (Required for new GUI)
PySide6>=6.2.0

# Scientific Computing & Visualization (Highly Recommended)
numpy>=1.19.0
matplotlib>=3.3.0

# Advanced Features (Recommended for full functionality)
scipy>=1.5.0
Pillow>=8.0.0

# Development tools (Optional)
# black>=22.0.0
# flake8>=4.0.0
# pytest>=7.0.0
# pytest-cov>=3.0.0
# sphinx>=4.5.0
